[Feature] Support for TDM bus arbitration from PECC
This commit has two main code changes. Both code changes are required to support the TDM bus interconnect for PECC and other predictable cache coherence mechanisms on newer gem5. The first set of code changes is the TDM bus logic contained (changes in ). The second set of code changes focuses on the build infrastructure. PECC was designed on a prior version of gem5 where the ruby build did not create separate directories for coherence protocols. In PECC, the bus logic uses some of the autogenerated structures defined in SLICC files, and assumes they are available in the same directory and namespace as the bus logic. In recent gem5 releases, the coherence logic is generated and contained in a separate directory and namespace. To support PECC, the build flow is modified to autogenerate files that the bus logic requires and move their location and adjust their namespaces.